+++ app/src/main/java/com/github/viktorpenelski/evtconf/spreadsheet/settings/SettingsEntryCache.kt
@@ -1,5 +1,7 @@
 package com.github.viktorpenelski.evtconf.spreadsheet.mainstage
 
+import com.github.viktorpenelski.evtconf.spreadsheet.ActiveSpreadsheetProperties
+import com.github.viktorpenelski.evtconf.spreadsheet.ActiveSpreadsheetProperties.Companion.PropertyTag.SETTINGS
 import com.github.viktorpenelski.evtconf.spreadsheet.SpreadsheetEntryCache
 import java.util.*
 
@@ -11,7 +13,6 @@ import java.util.*
  */
 class SettingsEntryCache private constructor() : SpreadsheetEntryCache<SettingsEntry> {
 
-    private val URL = "https://spreadsheets.google.com/feeds/list/1_Ol_0bP-S3GqEXEGPL3ODKmHAdWBBXcOdE3_M4phVe0/3/public/values?alt=json"
     private val factory = SettingsEntryFactory()
     private var cachedMainStageEntries = listOf<SettingsEntry>()
     private var lastCached = Calendar.getInstance()
@@ -40,7 +41,7 @@ class SettingsEntryCache private constructor() : SpreadsheetEntryCache<SettingsE
     }
 
     override fun getUrl(): String {
-        return URL
+        return ActiveSpreadsheetProperties.INSTANCE.getUrlFor(SETTINGS)
     }
 
     override fun getFactory(): SettingsEntryFactory {